modiy and delete

+++ week-3/PDO-harjotus/operations/selectData.php
@@ -1,8 +1,16 @@
 <?php
+if(session_status() != PHP_SESSION_ACTIVE) {
+    session_start();
+}
 global $DBH;
 global $SITE_URL;
 require_once __DIR__ . "/../config/config.php";
 require_once __DIR__ . '/../db/dbConnect.php';
+
+if (!isset($_SESSION['user'])) {
+    header('Location: '. $SITE_URL . '/user.php');
+    exit;
+}
 
 $sql = "SELECT MediaItems.*, Users.username FROM MediaItems JOIN Users ON Users.user_id = MediaItems.user_id;";
 
@@ -19,8 +27,14 @@ try {
             <td><img src="<?php echo $SITE_URL; ?>/uploads/<?php echo $row['filename']; ?>"
                      alt="<?php echo $row['title']; ?>"></td>
             <td>
-                <button class="modify-button" data-media_id="<?php echo $row['media_id']; ?>">Modify</button>
-                <a href="<?php echo $SITE_URL; ?>/operations/deleteData.php?media_id=<?php echo $row['media_id']; ?>">Delete</a>
+                <?php if ($_SESSION['user']['user_id'] == $row['user_id']
+                    || $_SESSION['user']['user_level_id'] == 1): ?>
+                    <button class="modify-button" data-media_id="<?php echo $row['media_id']; ?>">Modify</button>
+                    <!--suppress HtmlUnknownTarget -->
+                    <a href="<?php echo $SITE_URL; ?>/operations/deleteData.php?media_id=<?php echo $row['media_id']; ?>">Delete</a>
+                <?php else: ?>
+                    -
+                <?php endif; ?>
             </td>
         </tr>
     <?php
